fix: imports for restructured files
