+++ apps/journeys-admin/src/components/JourneyList/ArchivedJourneyList/ArchivedJourneyList.spec.tsx
@@ -4,9 +4,9 @@ import noop from 'lodash/noop'
 import { SnackbarProvider } from 'notistack'
 import { AuthUser } from 'next-firebase-auth'
 
-import { defaultJourney, oldJourney } from '../../journeyListData'
-import { ThemeProvider } from '../../../ThemeProvider'
-import { SortOrder } from '../../JourneySort'
+import { defaultJourney, oldJourney } from '../journeyListData'
+import { ThemeProvider } from '../../ThemeProvider'
+import { SortOrder } from '../JourneySort'
 import {
   ArchivedJourneyList,
   GET_ARCHIVED_JOURNEYS,
