+++ apps/journeys-admin/src/components/JourneyList/TrashedJourneyList/TrashedJourneyList.stories.tsx
@@ -1,13 +1,13 @@
 import { Story, Meta } from '@storybook/react'
 import { MockedProvider } from '@apollo/client/testing'
 import noop from 'lodash/noop'
-import { journeysAdminConfig } from '../../../../libs/storybook'
+import { journeysAdminConfig } from '../../../libs/storybook'
 import {
   defaultJourney,
   oldJourney,
   descriptiveJourney,
   publishedJourney
-} from '../../journeyListData'
+} from '../journeyListData'
 import { GET_TRASHED_JOURNEYS } from './TrashedJourneyList'
 import { TrashedJourneyList } from '.'
